feat: event processors rename and improvements
