+++ src/storage/index.ts
@@ -9,7 +9,7 @@ import { Application, CachedService } from '../definitions'
 import { loggingFactory } from '../logger'
 import { getEventsEmitterForService, isServiceInitialized } from '../blockchain/utils'
 import hooks from './storage.hooks'
-import eventProcessor from './storage.blockchain'
+import eventProcessor from './storage.processor'
 import Price from './models/price.model'
 import { confFactory } from '../conf'
 import { ethFactory } from '../blockchain'
